api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
fastapi-users[sqlalchemy]>=13.0.0
arq>=0.26.0
beautifulsoup4>=4.12.0
httpx>=0.28.0
lxml>=5.3.0
selectolax>=0.3.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
openai>=1.58.0
anthropic>=0.40.0
stripe>=8.0.0
tenacity>=9.0.0
structlog>=24.0.0
python-multipart>=0.0.18
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-genai>=1.0.0
Pillow>=10.0.0
